+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- hnefatafl-client: Allow going into fullscreen mode with F11.
+
 ## [6.1.1] - 2026-07-13
 
 ### Fixed